combine websockets js into single file
